pyproject.toml
[tool.poetry]
name = "instaauto"
version = "0.1.0"
description = ""
authors = ["Brandon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
appium-python-client = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
